Валидация формы на проверяемое лицо
